[ADD] User can view in detail selected notes for desktop

+++ frontend/src/hooks/noteCRUD.tsx
@@ -1,10 +1,12 @@
 export const deleteNote = async (id: number) => {
-  const response = await fetch(`/api/notes/${id}/`, {
-    method: "DELETE",
-    headers: {
-      'Content-Type': 'application/json'
-    },
-  })
-  const status = response.json()
-  return status.status
+  if (typeof id == 'number'){
+    const response = await fetch(`/api/notes/${id}/`, {
+      method: "DELETE",
+      headers: {
+        'Content-Type': 'application/json'
+      },
+    })
+    const status = response.json()
+    return status
+  }
 }
